Move TestingOpenLdapContainer to trino-testing-containers
