+++ plugin/trino-password-authenticators/src/test/java/io/trino/plugin/password/ldap/TestLdapAuthenticatorWithTimeouts.java
@@ -17,8 +17,9 @@ import com.google.common.io.Closer;
 import io.airlift.units.Duration;
 import io.trino.plugin.base.ldap.JdkLdapClient;
 import io.trino.plugin.base.ldap.LdapClientConfig;
-import io.trino.plugin.password.ldap.TestingOpenLdapServer.DisposableSubContext;
 import io.trino.spi.security.BasicPrincipal;
+import io.trino.testing.containers.TestingOpenLdapServer;
+import io.trino.testing.containers.TestingOpenLdapServer.DisposableSubContext;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
@@ -29,7 +30,7 @@ import org.testcontainers.containers.ToxiproxyContainer.ContainerProxy;
 import java.io.IOException;
 
 import static eu.rekawek.toxiproxy.model.ToxicDirection.DOWNSTREAM;
-import static io.trino.plugin.password.ldap.TestingOpenLdapServer.LDAP_PORT;
+import static io.trino.testing.containers.TestingOpenLdapServer.LDAP_PORT;
 import static java.lang.String.format;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
