+++ testing/trino-testing-containers/src/main/java/io/trino/testing/containers/ldap/LdapObjectDefinition.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.password.ldap;
+package io.trino.testing.containers.ldap;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
