+++ testing/trino-testing-containers/src/main/java/io/trino/testing/containers/ldap/LdapUtil.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.password.ldap;
+package io.trino.testing.containers.ldap;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -31,7 +31,7 @@ import java.util.Map;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
 
-final class LdapUtil
+public final class LdapUtil
 {
     public static final String MEMBER = "member";
